Calculate gene expression measure (E).

Calculate E values for every sequence in the given codonTable object.

E(cTobject, subsets = list(), ribosomal = FALSE, id_or_name2 = "1",
  alt.init = TRUE, stop.rm = FALSE, filtering = "none",
  len.threshold = 80)

# S4 method for codonTable
E(cTobject, subsets = list(), ribosomal = FALSE,
  id_or_name2 = "1", alt.init = TRUE, stop.rm = FALSE,
  filtering = "none", len.threshold = 80)

Arguments

cTobject
A codonTable object.

subsets
A (named) list of logical vectors, the length of each equal to getlen(cTobject), i.e. the number of sequences in the set, or character vectors (of any length) containing KEGG/eggNOG annotations, or codonTable objects (of any length).

ribosomal
Logical, if TRUE, CU statistic is also calculated against the average CU of the ribosomal genes in the sequence set.

id_or_name2
A single string that uniquely identifies the genetic code to extract. Should be one of the values in the id or name2 columns of GENETIC_CODE_TABLE.

alt.init
logical, whether to use alternative initiation codons. Default is TRUE.

stop.rm
Logical, whether to remove stop codons. Default is FALSE.

filtering
Character vector, one of c("none", "soft", "hard"). Specifies whether sequences shorther than some threshold value of length (in codons), len.threshold, should be excluded from calculations. If "none" (default), length of sequences is not checked, if "soft", a warrning is printed if there are shorter sequences, and if "hard", these sequences are excluded from calculation.

len.threshold
Optional numeric, specifying sequence length, in codons, used for filtering.

Value

A matrix with E values for every specified subset (subsets, self, ribosomal) in columns.

For definition of gene expression measure (E), see Karlin and Mrazek (2000).

Examples

# load example DNA sequences
exampledir <- system.file("extdata", package = "coRdon")
cT <- codonTable(readSet(exampledir))

# calculate E with respect to the CU
# of ribosomal genes among the example DNA sequences
em <- E(cT, ribosomal = TRUE)
head(em)

#>      ribosomal
#> [1,] 0.9946877
#> [2,] 1.0044203
#> [3,] 0.9998657
#> [4,] 0.9959171
#> [5,] 1.0056261
#> [6,] 1.0023404

# calculate E distance with respect to the average CU
# of the first 20 example DNA sequences
# (i.e. the first half of the example DNA set)
em <- E(cT, subsets = list(half = c(rep(TRUE, 20), rep(FALSE, 20))))

# alternatively, you can specify codonTable as a subset
halfcT <- codonTable(codonCounts(cT)[1:20,])
em2 <- E(cT, subsets = list(half = halfcT))
all.equal(em, em2) # TRUE

#> [1] TRUE

# filtering
E(cT, ribosomal = TRUE,
  filtering = "hard", len.threshold = 80) # E for 9 sequences (note that,

#>       ribosomal
#>  [1,]         1
#>  [2,]         1
#>  [3,]         1
#>  [4,]         1
#>  [5,]         1
#>  [6,]         1
#>  [7,]         1
#>  [8,]         1
#>  [9,]         1

                                          # accidentally, all are ribosomal)
sum(getlen(cT) > 80) # 9 sequences are longer than 80 codons

#> [1] 9

em1 <- E(cT, ribosomal = TRUE, filtering = "none") # no filtering
em2 <- E(cT, ribosomal = TRUE, filtering = "soft") # warning

#> Warning: Some sequences have below-threshold length!

all.equal(em1, em2) # TRUE

#> [1] TRUE

# options for genetic code
em <- E(cT, ribosomal = TRUE,
        stop.rm = TRUE) # don't use stop codons in calculation
em <- E(cT, ribosomal = TRUE,
        alt.init = FALSE) # don't use alternative start codons
em <- E(cT, ribosomal = TRUE,
        id_or_name2 = "2") # use different genetic code, for help
                           # see `?Biostrings::GENETIC_CODE`