Calculate eorderliness of synonymous codon usage (SCUO).

Calculate SCUO values for every sequence in the given codonTable object.

SCUO(cTobject, id_or_name2 = "1", alt.init = TRUE, stop.rm = FALSE,
  filtering = "none", len.threshold = 80)

# S4 method for codonTable
SCUO(cTobject, id_or_name2 = "1", alt.init = TRUE,
  stop.rm = FALSE, filtering = "none", len.threshold = 80)

Arguments

cTobject
A codonTable object.

id_or_name2
A single string that uniquely identifies the genetic code to extract. Should be one of the values in the id or name2 columns of GENETIC_CODE_TABLE.

alt.init
logical, whether to use alternative initiation codons. Default is TRUE.

stop.rm
Logical, whether to remove stop codons. Default is FALSE.

filtering
Character vector, one of c("none", "soft", "hard"). Specifies whether sequences shorther than some threshold value of length (in codons), len.threshold, should be excluded from calculations. If "none" (default), length of sequences is not checked, if "soft", a warrning is printed if there are shorter sequences, and if "hard", these sequences are excluded from calculation.

len.threshold
Optional numeric, specifying sequence length, in codons, used for filtering.

Value

A numeric vector with SCUO values.

For definition of synonymous codon usage eorderliness (SCUO), see Wan et al. (2004).

Examples

# load example DNA sequences
exampledir <- system.file("extdata", package = "coRdon")
cT <- codonTable(readSet(exampledir))

# calculate SCUO
scuo <- SCUO(cT)
head(scuo)

#> [1] 0.4042227 0.3770169 0.2820085 0.4152137 0.3199568 0.4828782

# filtering
SCUO(cT, filtering = "hard", len.threshold = 80) # SCUO for 9 sequences

#> [1] 0.4042227 0.2773462 0.3467498 0.2084217 0.3167224 0.2650509 0.3741477
#> [8] 0.3357457 0.3561581

sum(getlen(cT) > 80) # 9 sequences are longer than 80 codons

#> [1] 9

scuo1 <- SCUO(cT, filtering = "none") # no filtering
scuo2 <- SCUO(cT, filtering = "soft") # warning

#> Warning: Some sequences have below-threshold length!

all.equal(scuo1, scuo2) # TRUE

#> [1] TRUE

# options for genetic code
scuo <- SCUO(cT, stop.rm = TRUE) # don't use stop codons in calculation
scuo <- SCUO(cT, alt.init = FALSE) # don't use alternative start codons
scuo <- SCUO(cT, id_or_name2 = "2") # use different genetic code, for help
                                    # see `?Biostrings::GENETIC_CODE`