An S4 class codonTable

Contains codon counts and optional annotation for a set DNA sequences.

codonTable(x)

# S4 method for DNAStringSet
codonTable(x)

# S4 method for matrix
codonTable(x)

# S4 method for data.frame
codonTable(x)

codonCounts(object)

# S4 method for codonTable
codonCounts(object)

getID(object)

# S4 method for codonTable
getID(object)

getlen(object)

# S4 method for codonTable
getlen(object)

getKO(object)

# S4 method for codonTable
getKO(object)

setKO(object, ann)

# S4 method for codonTable
setKO(object, ann)

getCOG(object)

# S4 method for codonTable
getCOG(object)

setCOG(object, ann)

# S4 method for codonTable
setCOG(object, ann)

Arguments

x
An object of DNAStringSet, matrix or data.frame class.

object
A codonTable object.

ann
A character vector of sequence annotations, must be of length equal to length(object).

Value

A codonTable.

Methods (by generic)

codonTable: Create new objects of class codonTable.

codonCounts: Get codon counts from codonTable object.

getID: Get IDs for codonTable class.

getlen: Get lengths of sequences in codonTable object.

getKO: Get KO annotations of sequences in codonTable object.

setKO: Set KO annotations for codonTable object.

getCOG: Get COG annotations of sequences in codonTable object.

setCOG: Set COG annotations for codonTable object.

Slots

ID

A character vector of sequence identifiers.

counts

A matrix containing codon counts. Columns are codons, rows are sequences.

len

A numeric vector,length equal to nrow(counts), containing lengths of sequnces.

KO

A character vector of KEGG annotations for sequences, length equal to nrow(counts). If no annotation is available, this will be an empty vector.

COG

A character vector of COG annotations for sequences, length equal to nrow(counts). If no annotation is available, this will be an empty vector.

Examples

# create codonTable with codon counts for sequences in DNAStringSet
require(Biostrings)

#> Loading required package: Biostrings

#> Loading required package: BiocGenerics

#> Loading required package: parallel

#> 
#> Attaching package: 'BiocGenerics'

#> The following objects are masked from 'package:parallel':
#> 
#>     clusterApply, clusterApplyLB, clusterCall, clusterEvalQ,
#>     clusterExport, clusterMap, parApply, parCapply, parLapply,
#>     parLapplyLB, parRapply, parSapply, parSapplyLB

#> The following objects are masked from 'package:stats':
#> 
#>     IQR, mad, sd, var, xtabs

#> The following objects are masked from 'package:base':
#> 
#>     anyDuplicated, append, as.data.frame, basename, cbind, colMeans,
#>     colnames, colSums, dirname, do.call, duplicated, eval, evalq,
#>     Filter, Find, get, grep, grepl, intersect, is.unsorted, lapply,
#>     lengths, Map, mapply, match, mget, order, paste, pmax, pmax.int,
#>     pmin, pmin.int, Position, rank, rbind, Reduce, rowMeans, rownames,
#>     rowSums, sapply, setdiff, sort, table, tapply, union, unique,
#>     unsplit, which, which.max, which.min

#> Loading required package: S4Vectors

#> Loading required package: stats4

#> 
#> Attaching package: 'S4Vectors'

#> The following object is masked from 'package:base':
#> 
#>     expand.grid

#> Loading required package: IRanges

#> 
#> Attaching package: 'IRanges'

#> The following object is masked from 'package:grDevices':
#> 
#>     windows

#> Loading required package: XVector

#> 
#> Attaching package: 'Biostrings'

#> The following object is masked from 'package:base':
#> 
#>     strsplit

dna <- DNAStringSet(c("ACGAAGTGTACTGTAATTTGCACAGTACTTAAATGT",
                      "ACGTCCGTACTGATCGATTCCGTGATT"))
cT <- codonTable(dna)
codonCounts(cT)

#>      AAA AAC AAG AAT ACA ACC ACG ACT AGA AGC AGG AGT ATA ATC ATG ATT CAA CAC
#> [1,]   1   0   1   0   1   0   1   1   0   0   0   0   0   0   0   1   0   0
#> [2,]   0   0   0   0   0   0   1   0   0   0   0   0   0   1   0   1   0   0
#>      CAG CAT CCA CCC CCG CCT CGA CGC CGG CGT CTA CTC CTG CTT GAA GAC GAG GAT
#> [1,]   0   0   0   0   0   0   0   0   0   0   0   0   0   1   0   0   0   0
#> [2,]   0   0   0   0   0   0   0   0   0   0   0   0   1   0   0   0   0   1
#>      GCA GCC GCG GCT GGA GGC GGG GGT GTA GTC GTG GTT TAA TAC TAG TAT TCA TCC
#> [1,]   0   0   0   0   0   0   0   0   2   0   0   0   0   0   0   0   0   0
#> [2,]   0   0   0   0   0   0   0   0   1   0   1   0   0   0   0   0   0   2
#>      TCG TCT TGA TGC TGG TGT TTA TTC TTG TTT
#> [1,]   0   0   0   1   0   2   0   0   0   0
#> [2,]   0   0   0   0   0   0   0   0   0   0

getlen(cT)

#> [1] 12  9

getKO(cT)

#> character(0)

cT <- setKO(cT, c("K00001", "K00002"))
getKO(cT)

#> [1] "K00001" "K00002"

# convert matrix containing codon counts to codonTable
mat <- matrix(sample(1:10, 122, replace = TRUE), nrow = 2)
codonTable(mat) # produces informative warning

#> Assigned alphabetically sorted codons as column names.

#> codonTable instance with codon counts from 2 sequences.
#>   length AAA AAC AAG AAT ACA ACC ACG ACT AGA AGC AGG AGT ATA ATC ATG ATT CAA
#> 1    330   1   7   1   5   8   9   1   5   5   4   3   8  10   1   7   1   4
#> 2    338   9   2   5   3   8   2   4   2   1  10   7   2   8   6   7   3   7
#>   CAC CAG CAT CCA CCC CCG CCT CGA CGC CGG CGT CTA CTC CTG CTT GAA GAC GAG GAT
#> 1   5   8   2   7   7   1   8  10   5   2   5   3   1   9   6   6  10   6   5
#> 2   5  10   3   5   7   8  10   4   4   6   8   8   4   3   4   2   6   3   4
#>   GCA GCC GCG GCT GGA GGC GGG GGT GTA GTC GTG GTT TAC TAT TCA TCC TCG TCT TGC
#> 1   1   4   4   9   6   3   5   5   7  10   6   8   6   1   9   5   5   6   6
#> 2   5   1   6   4   7   3  10   3   1   2   9   3   9   5   9   3   7   7   9
#>   TGG TGT TTA TTC TTG TTT
#> 1   8   1  10   4   6   9
#> 2   9   9   6   8   7   6