An S4 class crossTab

Contingency table of sequences' annotations and the corresponding numeric values.

crossTab(sequences, variable, threshold = 1L, percentiles = NULL)

# S4 method for character,numeric
crossTab(sequences, variable, threshold = 1L,
  percentiles = NULL)

getSeqAnnot(x)

# S4 method for crossTab
getSeqAnnot(x)

getVariable(x)

# S4 method for crossTab
getVariable(x)

contable(x)

# S4 method for crossTab
contable(x)

Arguments

sequences
Character vector of sequences' annotations (KO, COG).

variable
Numeric vector of the coresponding CU values.

threshold
A threshold value (or a vector of values) of the variable. Sequences with value of the given variable greater than threshold are taken as a subset. Default is 1. If no threshold should be set, specify threshold = NULL

percentiles
A single value or a vector of values between 0 and 1. Sequences with value of the given variable in the top percentiles are taken as a subset. If no percentiles should be specified, the argument takes the value NULL.

x
A crossTab object.

Value

Returns a crossTab object with category values in rows, and with separate columns for counts in background (all) and subsets, i.e. for diferrent thresholds/percentiles provided.

Methods (by generic)

crossTab: Create a contingency table for the set of annotated sequences and the corresponding codon usage (CU) values.

getSeqAnnot: Get sequence annotations from crossTab object.

getVariable: Get values of the variable used to create contingency table in crossTab object.

contable: Get contingency table from crossTab object.

Slots

sequences

Character vector of sequences annotations.

variable

Numeric vector of the coresponding CU values.

table

Contingecy table.

Examples

set.seed(5491)
s <- sample(LETTERS[1:3], 10, replace = TRUE)
v <- sample(1:5, 10, replace = TRUE)
crossTab(s, v)

#> crossTab instance for 10 sequences.
#> Sequence annotations:
#>   chr [1:10] "B" "C" "C" "C" "B" "A" "B" "A" "A" "B" 
#> Corresponding variable values:
#>   int [1:10] 3 2 4 1 2 1 3 2 2 4 
#> Contingecy table:
#>    category all gt_1
#> 1:        A   3    3
#> 2:        B   4    4
#> 3:        C   3    3

crossTab(s, v, threshold = c(3,5))

#> crossTab instance for 10 sequences.
#> Sequence annotations:
#>   chr [1:10] "B" "C" "C" "C" "B" "A" "B" "A" "A" "B" 
#> Corresponding variable values:
#>   int [1:10] 3 2 4 1 2 1 3 2 2 4 
#> Contingecy table:
#>    category all gt_3 gt_5
#> 1:        A   3    0    0
#> 2:        B   4    3    0
#> 3:        C   3    1    0

crossTab(s, v, threshold = NULL, percentiles = c(0.5, 0.3))

#> crossTab instance for 10 sequences.
#> Sequence annotations:
#>   chr [1:10] "B" "C" "C" "C" "B" "A" "B" "A" "A" "B" 
#> Corresponding variable values:
#>   int [1:10] 3 2 4 1 2 1 3 2 2 4 
#> Contingecy table:
#>    category all top_0.5 top_0.3
#> 1:        A   3       2       0
#> 2:        B   4       4       3
#> 3:        C   3       2       1

ct <- crossTab(s, v)
contable(ct)

#>    category all gt_1
#> 1:        A   3    3
#> 2:        B   4    4
#> 3:        C   3    3

getSeqAnnot(ct)

#>  [1] "B" "C" "C" "C" "B" "A" "B" "A" "A" "B"

getVariable(ct)

#>  [1] 3 2 4 1 2 1 3 2 2 4