Extract chosen enrichment values to a matrix.

Extract enrichment values from multiple samples, i.e. AnnotatedDataFrame objects. Note that the samples should contain annotations of the same type (i.e. the same ontology). The data in matrix format can be easily used in different types of downstream analyses, such as GAGE, and visualised, e.g. using a heatmap.

enrich.matrix(x, variable, replace.na = TRUE)

# S4 method for list
enrich.matrix(x, variable, replace.na = TRUE)

Arguments

x
A named list of AnnotatedDataFrame objects.

variable
Character, indicating the statistic values to extract from enrich.data.frame objects in x, must be one of c("enrich","M","A").

replace.na
logical, whether to replace NA values in the output. If `TRUE` (default), NAs will be replaced by 0. Alternatively, if numueric, NAs will be replaced by that given value.

Value

matrix with sequences' annotations as rows, and variable values for different samples as columns.

Examples

require(Biobase)

#> Loading required package: Biobase

#> Welcome to Bioconductor
#> 
#>     Vignettes contain introductory material; view with
#>     'browseVignettes()'. To cite Bioconductor, see
#>     'citation("Biobase")', and for packages 'citation("pkgname")'.

# create contingency table
s <- getKO(LD94)
v <- as.numeric(MELP(LD94, ribosomal = TRUE))
ct <- crossTab(s, v, percentiles = 0.2)

# enrichment analysis
enr <- enrichment(ct)
enr # for help, see `?Biobase::AnnotatedDataFrame`

#> $top_0.2
#> An object of class 'AnnotatedDataFrame'
#>   rowNames: 1 2 ... 607 (607 total)
#>   varLabels: category all ... padj (8 total)
#>   varMetadata: labelDescription
#> 
#> $gt_1
#> An object of class 'AnnotatedDataFrame'
#>   rowNames: 1 2 ... 607 (607 total)
#>   varLabels: category all ... padj (8 total)
#>   varMetadata: labelDescription
#> 

head(pData(enr$top_0.2), 10)

#>    category all top_0.2    enrich          M         A     pvals      padj
#> 1    K00003   1       0 -16.66667 -0.2630344 0.1315172 1.0000000 1.0000000
#> 2    K00012   1       0 -16.66667 -0.2630344 0.1315172 1.0000000 1.0000000
#> 3    K00016   1       0 -16.66667 -0.2630344 0.1315172 1.0000000 1.0000000
#> 4    K00027   1       0 -16.66667 -0.2630344 0.1315172 1.0000000 1.0000000
#> 5    K00033   1       1  66.66667  0.7369656 0.6315172 0.1813512 0.9489669
#> 6    K00040   1       1  66.66667  0.7369656 0.6315172 0.1813512 0.9489669
#> 7    K00048   1       0 -16.66667 -0.2630344 0.1315172 1.0000000 1.0000000
#> 8    K00052   3       0 -37.50000 -0.6780719 0.3390360 1.0000000 1.0000000
#> 9    K00057   1       0 -16.66667 -0.2630344 0.1315172 1.0000000 1.0000000
#> 10   K00058   1       0 -16.66667 -0.2630344 0.1315172 1.0000000 1.0000000

head(pData(enr$gt_1), 10)

#>    category all gt_1    enrich          M          A     pvals      padj
#> 1    K00003   1    0 -10.71429 -0.1634987 0.08174937 1.0000000 1.0000000
#> 2    K00012   1    0 -10.71429 -0.1634987 0.08174937 1.0000000 1.0000000
#> 3    K00016   1    0 -10.71429 -0.1634987 0.08174937 1.0000000 1.0000000
#> 4    K00027   1    0 -10.71429 -0.1634987 0.08174937 1.0000000 1.0000000
#> 5    K00033   1    0 -10.71429 -0.1634987 0.08174937 1.0000000 1.0000000
#> 6    K00040   1    1  78.57143  0.8365013 0.58174937 0.1131328 0.9952408
#> 7    K00048   1    0 -10.71429 -0.1634987 0.08174937 1.0000000 1.0000000
#> 8    K00052   3    0 -26.47059 -0.4436067 0.22180333 1.0000000 1.0000000
#> 9    K00057   1    0 -10.71429 -0.1634987 0.08174937 1.0000000 1.0000000
#> 10   K00058   1    0 -10.71429 -0.1634987 0.08174937 1.0000000 1.0000000

enrm <- enrich.matrix(enr, "M")
head(enrm)

#>           top_0.2       gt_1
#> K00003 -0.2630344 -0.1634987
#> K00012 -0.2630344 -0.1634987
#> K00016 -0.2630344 -0.1634987
#> K00027 -0.2630344 -0.1634987
#> K00033  0.7369656 -0.1634987
#> K00040  0.7369656  0.8365013