Enrichment analysis for codon usage (CU) data.

Performs enrichment analysis, given a contongency table of codon counts. p values are calculated by binomial test, adjustment for multiple testing can be performed by any of the p.adjust.methods.

enrichment(x, pvalueCutoff = numeric(), pAdjustMethod = "BH",
  padjCutoff = numeric())

# S4 method for crossTab
enrichment(x, pvalueCutoff = numeric(),
  pAdjustMethod = "BH", padjCutoff = numeric())

Arguments

x
A crossTab object

pvalueCutoff
Numeric, discard categories with p value below this threshold. By default, no threshold is set (numeric()).

pAdjustMethod
Character, one of the p.adjust.methods.

padjCutoff
Numeric, discard categories with adjusted p value below this threshold. By default, no threshold is set (numeric()).

Value

An AnnotatedDataFrame object, or a list of those; data in each object has category values in rows, and the following columns:

category, a character vector of annotation categories

all, a numeric vector of integers, coresponding to sequence counts for each annotation category, in the background gene set (universe).

a numeric vector(s) of integers, coresponding to sequence counts for each annotation category, in the set of genes for which enrichment is calculated, i.e. the predefined subset of (usually highly expressed) genes in the universe (named for the corresponding `crossTab` column).

enrichment, calculated as the ratio: (scaled sample counts - scaled backg. counts) / scaled backg. counts * 100, where scaling means that sample counts are simply increased by 1, and background counts are multiplied by ratio of summed sample counts and summed backgroun counts, and also increased by 1

M, log ratios of scaled counts

A, mean average of scaled counts

pvals, p values for exact binomial test

padj, p values corrected by BH method.

Examples

require(Biobase)

# create contingency table
s <- getKO(HD59)
v <- as.numeric(MELP(HD59, ribosomal = TRUE))
ct <- crossTab(s, v)

# enrichment analysis
enr <- enrichment(ct)
enr # for help, see `?Biobase::AnnotatedDataFrame`

#> An object of class 'AnnotatedDataFrame'
#>   rowNames: 1 2 ... 618 (618 total)
#>   varLabels: category all ... padj (8 total)
#>   varMetadata: labelDescription

head(pData(enr))

#>   category all gt_1    enrich          M         A      pvals padj
#> 1   K00003   1    0 -13.49481 -0.2091414 0.1045707 1.00000000    1
#> 2   K00010   1    0 -13.49481 -0.2091414 0.1045707 1.00000000    1
#> 3   K00012   4    1  23.15271  0.3004484 0.8497758 0.46487309    1
#> 4   K00014   2    0 -23.78049 -0.3917677 0.1958839 1.00000000    1
#> 5   K00024   3    2 104.35967  1.0311105 1.0694072 0.08042498    1
#> 6   K00029   1    0 -13.49481 -0.2091414 0.1045707 1.00000000    1

enr <- enrichment(ct, pAdjustMethod = "holm")
head(pData(enr))

#>   category all gt_1    enrich          M         A      pvals padj
#> 1   K00003   1    0 -13.49481 -0.2091414 0.1045707 1.00000000    1
#> 2   K00010   1    0 -13.49481 -0.2091414 0.1045707 1.00000000    1
#> 3   K00012   4    1  23.15271  0.3004484 0.8497758 0.46487309    1
#> 4   K00014   2    0 -23.78049 -0.3917677 0.1958839 1.00000000    1
#> 5   K00024   3    2 104.35967  1.0311105 1.0694072 0.08042498    1
#> 6   K00029   1    0 -13.49481 -0.2091414 0.1045707 1.00000000    1

enr <- enrichment(ct, pvalueCutoff = 0.05)
head(pData(enr))

#>   category all gt_1   enrich        M         A      pvals padj
#> 1   K00257   2    2 128.6585 1.193195 0.9883651 0.03948269    1
#> 2   K00262   2    2 128.6585 1.193195 0.9883651 0.03948269    1
#> 3   K00611   3    3 172.4796 1.446148 1.2769260 0.01191984    1
#> 4   K00948   2    2 128.6585 1.193195 0.9883651 0.03948269    1
#> 5   K01006   3    3 172.4796 1.446148 1.2769260 0.01191984    1
#> 6   K01610   2    2 128.6585 1.193195 0.9883651 0.03948269    1

enr <- enrichment(ct, padjCutoff = 0.05)
head(pData(enr))

#> [1] category all      gt_1     enrich   M        A        pvals    padj    
#> <0 rows> (or 0-length row.names)