An S4 class genCode

Object of genCode class describes the variant of genetic code to be used in CU calculations.

genCode(id_or_name2 = "1", alt.init = TRUE, stop.rm = FALSE)

# S4 method for ANY
genCode(id_or_name2 = "1", alt.init = TRUE,
  stop.rm = FALSE)

Arguments

id_or_name2
A single string that uniquely identifies the genetic code to extract. Should be one of the values in the id or name2 columns of GENETIC_CODE_TABLE.

alt.init
logical, whether to use alternative initiation codons. Default is TRUE.

stop.rm
logical, whether to remove stop codons. Default is FALSE.

Value

A genCode object.

Methods (by generic)

genCode: Creates new instances of genCode class.

Slots

ctab

A data.table with two colums: codon and AA, amino acid.

codons

A character vector of codons.

stops

A character vector of stop codons. Note that, if stop.rm is TRUE, this will be an empty vector.

nostops

A character vector of no-stop codons. If stop.rm is TRUE, this will be equal to the codons slot.

cl

A list, each element of which is a vector of integers indicating the positions of synonymous codons for that amino acid, when codons are alphabetically ordered.

deg

A numeric vector of degeneracies for alphabetically ordered amino acids.