Read set of sequences.

Reads a set of fasta files stored in folder, or a single fasta file.

readSet(folder = character(), file = character(), KOs = c(),
  zipped = FALSE)

Arguments

folder
Path to directory containing .fasta files.

file
Path to a single .fasta file, or zipped file (if latter, specify ZIPPED = TRUE).

KOs
An optional character vector of sequence annotations (e.g. KO) contained in the names of fasta files to be selectively read.

zipped
Logical, whether folder or file is zipped. Default is FALSE.

Value

Returns a DNAStringSet object.

Examples

exampledir <- system.file("extdata", package = "coRdon")
files <- list.files(exampledir)
readSet(folder = exampledir)

#>   A DNAStringSet instance of length 40
#>      width seq                                              names               
#>  [1]   264 TTGAGTAGACTTAGAGAACAGTA...AGGAATGCCAATCGGATGTAAG contigs-K02931.fa...
#>  [2]   195 AGAGGTGTTAACCCTAACGCATT...GTTCAATATGCCATTCGCTAAA contigs-K02931.fa...
#>  [3]   153 GTGAGCAGACTGAAAGAACAGTA...AGAGAACGCAAAGTTATTAGAG contigs-K02931.fa...
#>  [4]   123 CCGGAAATCGAGTACGATAAGGT...GTTCAAAATGCCGTTTACAAAA contigs-K02931.fa...
#>  [5]   201 GGTACCGTTAAGAATATCACTTC...TCTCGGCTTGAAACAACTCACT contigs-K02945.fa...
#>  ...   ... ...
#> [36]   183 ATTCCATTGCCTACACACAAGCG...TAGCGGTGTAGAAGTAGAAATT contigs-K02946.fa...
#> [37]   162 CCTCTGCCTACGCACAAGCGTAT...GATGAAGTTGGAGTTACCGAGT contigs-K02946.fa...
#> [38]   165 CTTCCTACGCACAAACGTATCTT...GTTGGAGTTGCCGAGTGNTGTG contigs-K02946.fa...
#> [39]   159 ACCGTAAACCGTTCGACTTTCGT...TGTGGAAGTAGAAATTAAAGTT contigs-K02946.fa...
#> [40]   135 ATGCTGGGTGCTTTTTTCATTTC...AATTAGAAATGAGGTGAAGTCA contigs-K03496.fa...

readSet(folder = exampledir, KOs = "K02931")

#>   A DNAStringSet instance of length 4
#>     width seq                                               names               
#> [1]   264 TTGAGTAGACTTAGAGAACAGTA...AAGGAATGCCAATCGGATGTAAG contigs-K02931.fa...
#> [2]   195 AGAGGTGTTAACCCTAACGCATT...AGTTCAATATGCCATTCGCTAAA contigs-K02931.fa...
#> [3]   153 GTGAGCAGACTGAAAGAACAGTA...AAGAGAACGCAAAGTTATTAGAG contigs-K02931.fa...
#> [4]   123 CCGGAAATCGAGTACGATAAGGT...AGTTCAAAATGCCGTTTACAAAA contigs-K02931.fa...

pathtofile <- paste(exampledir, files[1], sep = "/")
readSet(file = pathtofile)

#>   A DNAStringSet instance of length 4
#>     width seq                                               names               
#> [1]   264 TTGAGTAGACTTAGAGAACAGTA...AAGGAATGCCAATCGGATGTAAG contigs-K02931.fa...
#> [2]   195 AGAGGTGTTAACCCTAACGCATT...AGTTCAATATGCCATTCGCTAAA contigs-K02931.fa...
#> [3]   153 GTGAGCAGACTGAAAGAACAGTA...AAGAGAACGCAAAGTTATTAGAG contigs-K02931.fa...
#> [4]   123 CCGGAAATCGAGTACGATAAGGT...AGTTCAAAATGCCGTTTACAAAA contigs-K02931.fa...