Reduce crossTab.

Reduce the input contingency table by associating sequences with KEGG Pathway, KEGG Module or COG functional category identifiers.

reduceCrossTab(x, target)

# S4 method for crossTab,character
reduceCrossTab(x, target)

Arguments

x
A crossTab object to be reduced.

target
Character vector indicating which onthology to use, either "pathway" or "module", or "cogfunction".

Value

Returns input crossTab object, with updated contingency table, displaying new category values in rows, and updated counts in columns.

Examples

# create contingency table
s <- getKO(HD59)
v <- as.numeric(MELP(HD59, ribosomal = TRUE))
ct <- crossTab(s, v)
ct

#> crossTab instance for 1000 sequences.
#> Sequence annotations:
#>   chr [1:1000] "K01875" "K02190" "K11131" "K01791" "K00754" "K01883" ... 
#> Corresponding variable values:
#>   num [1:1000] 0.965 1.261 0.791 0.887 0.595 ... 
#> Contingecy table:
#>      category all gt_1
#>   1:   K00003   1    0
#>   2:   K00010   1    0
#>   3:   K00012   4    1
#>   4:   K00014   2    0
#>   5:   K00024   3    2
#>  ---                  
#> 614:   K13643   1    0
#> 615:   K13652   1    0
#> 616:   K13668   1    0
#> 617:   K13694   1    0
#> 618:   K13747   2    0

# reduce contingency table
reduceCrossTab(ct, "pathway")

#> crossTab instance for 1000 sequences.
#> Sequence annotations:
#>   chr [1:1000] "K01875" "K02190" "K11131" "K01791" "K00754" "K01883" ... 
#> Corresponding variable values:
#>   num [1:1000] 0.965 1.261 0.791 0.887 0.595 ... 
#> Contingecy table:
#>      category all gt_1
#>   1: map00010  21    6
#>   2: map00020  15    7
#>   3: map00030  19    4
#>   4: map00040  17    4
#>   5: map00051  13    3
#>  ---                  
#> 180: map05206   2    1
#> 181: map05215   1    0
#> 182: map05230   4    1
#> 183: map05340   2    0
#> 184: map05418   3    0

reduceCrossTab(ct, "module")

#> crossTab instance for 1000 sequences.
#> Sequence annotations:
#>   chr [1:1000] "K01875" "K02190" "K11131" "K01791" "K00754" "K01883" ... 
#> Corresponding variable values:
#>   num [1:1000] 0.965 1.261 0.791 0.887 0.595 ... 
#> Contingecy table:
#>      category all gt_1
#>   1:   M00001  10    3
#>   2:   M00002   4    2
#>   3:   M00003  10    5
#>   4:   M00004   7    1
#>   5:   M00005   2    2
#>  ---                  
#> 199:   M00793   5    1
#> 200:   M00821   2    0
#> 201:   M00841   4    0
#> 202:   M00844   5    3
#> 203:   M00845   4    0