Subset codonTable object.

Subset codonTable object.

# S3 method for codonTable
subset(x, subset, ...)

Arguments

x
A codonTable object to be subset.

subset
A logical or character vector indicating which elements of x to keep. If logical, subset should be of length length(x). If character, subset should contain at least some of the elements of either getKO(x) or getCOG(x).

...
further arguments to be passed to or from other methods.

Value

subsets of codonTable object, keeping in each slot only those elements that meet the criteria in subset.

Examples

# create codonTable
mat <- matrix(sample(1:10, 610, replace = TRUE), nrow = 10)
cT <- codonTable(mat) # produces informative warning

#> Assigned alphabetically sorted codons as column names.

cT

#> codonTable instance with codon counts from 10 sequences.
#>    length AAA AAC AAG AAT ACA ACC ACG ACT AGA AGC AGG AGT ATA ATC ATG ATT CAA
#> 1     357  10   9  10   4   1   2   8   7   8   9   9   1   6   3   5   8   4
#> 2     353   8   2   8   3   7   2   3   1   7   8   4   9   7   5   9   8   4
#> 3     327   3   1   2   3   9   1   3   8   8   9   5   8   5   4   2   1   5
#> 4     360   2   5   1   8   6   6   2   9  10   6   3   3   6   6   1   4   8
#> 5     299   9   8   2   3   1   3   3   2   6   8   7   1   1   5   6   3   4
#> 6     310   8   2   7   1   2   6   7   6   8  10   1   6   7   6   5  10   6
#> 7     344   2   7   9   1   2   9   8   8   8   5   8  10   7   2   7  10   3
#> 8     339   8   5  10   4   4   1   3   9   6  10   7   6   4   4   2   7   1
#> 9     307   4   5   1   1   7   3   6   6   9   2   6  10   9   7  10   9   5
#> 10    309   7   4   5   4  10   1   6   2   9   1   3  10  10   1   8   1   6
#>    CAC CAG CAT CCA CCC CCG CCT CGA CGC CGG CGT CTA CTC CTG CTT GAA GAC GAG GAT
#> 1   10   6   3   5   4   5   4   6   2   4   9   4   1  10   4   9   9   5   2
#> 2   10   2   9   2   3   4   8   8  10   6   7   1   9   9   8  10   3   3   7
#> 3    5   1   7   1   3   8   2   9   1   4  10   7   7   5   6   3   4   9   2
#> 4   10   9   1   7   9   8   4   1   9  10   9   8   9   5   6   8  10   9   8
#> 5    5   3   1   8   1  10   9   3   8   8   7   3   1   7   6   1   2   7   8
#> 6    3   1   1   7   8  10   3   6   1   3   8   7   3   3   3   5   2   1   4
#> 7    4   7   4   4  10   8   2   2   9   3   3   4   9   2   3  10   8   3  10
#> 8    2  10   4   8   3   2   7   7   4  10   1   1  10   4   8   3   2   7   6
#> 9    3   1  10   1   5   2   9   2   8   7   4   7   2   5   1  10   7   6   8
#> 10   2   5   4   8   3   2   5   7   4   3   1   6   5   5  10   1  10   1   2
#>    GCA GCC GCG GCT GGA GGC GGG GGT GTA GTC GTG GTT TAC TAT TCA TCC TCG TCT TGC
#> 1    9  10   7   2   4   1   3   2   8   5   9  10   9   5   5   8   4   6  10
#> 2    1   9   7   1   5  10   9   3   1  10   4   1   4   3   9   2  10   9  10
#> 3    8   3   1   6   7  10   5   9   7   1   3   7   8   5   6   1  10   8   9
#> 4    7  10   7   1   6   5   1   7   3   3   7   4  10   8   7   2   1   6   9
#> 5    8   8   8   3   4   8   5   3   3   1   1   2   6   2   8  10   6  10   4
#> 6    3   2   4   7   8  10   8   2   3   3   2   3   2   3  10   7   8   1   7
#> 7   10   5   1   8   4   5   4   4   4   6   4   7   2   8   4   7   7   1   4
#> 8    5  10   5   2   1   6  10   9   6   4  10   3   2   4   5  10   6   9   9
#> 9    1   8   5  10   2   4   4   1   9   2   5   8   8   2   8   1   3   5   3
#> 10   3   8   6   2   9   3   5   4   7   2   8   7   5   2   5   9   1   5  10
#>    TGG TGT TTA TTC TTG TTT
#> 1    2   5   3   8   6  10
#> 2    3   6   3  10   1   8
#> 3    8   3   5  10   7   9
#> 4    3   3   6   6   8   4
#> 5    4  10   2   1   7   5
#> 6   10   3  10   1   8   8
#> 7    4   5   4   5  10  10
#> 8    8   2   5   7   7   4
#> 9    7   2   1   2   7   1
#> 10   8   5   5   5   7   6

subset(cT, c(rep(c(TRUE,FALSE), 5))) # subset odd sequences

#> codonTable instance with codon counts from 5 sequences.
#>   length AAA AAC AAG AAT ACA ACC ACG ACT AGA AGC AGG AGT ATA ATC ATG ATT CAA
#> 1    357  10   9  10   4   1   2   8   7   8   9   9   1   6   3   5   8   4
#> 2    327   3   1   2   3   9   1   3   8   8   9   5   8   5   4   2   1   5
#> 3    299   9   8   2   3   1   3   3   2   6   8   7   1   1   5   6   3   4
#> 4    344   2   7   9   1   2   9   8   8   8   5   8  10   7   2   7  10   3
#> 5    307   4   5   1   1   7   3   6   6   9   2   6  10   9   7  10   9   5
#>   CAC CAG CAT CCA CCC CCG CCT CGA CGC CGG CGT CTA CTC CTG CTT GAA GAC GAG GAT
#> 1  10   6   3   5   4   5   4   6   2   4   9   4   1  10   4   9   9   5   2
#> 2   5   1   7   1   3   8   2   9   1   4  10   7   7   5   6   3   4   9   2
#> 3   5   3   1   8   1  10   9   3   8   8   7   3   1   7   6   1   2   7   8
#> 4   4   7   4   4  10   8   2   2   9   3   3   4   9   2   3  10   8   3  10
#> 5   3   1  10   1   5   2   9   2   8   7   4   7   2   5   1  10   7   6   8
#>   GCA GCC GCG GCT GGA GGC GGG GGT GTA GTC GTG GTT TAC TAT TCA TCC TCG TCT TGC
#> 1   9  10   7   2   4   1   3   2   8   5   9  10   9   5   5   8   4   6  10
#> 2   8   3   1   6   7  10   5   9   7   1   3   7   8   5   6   1  10   8   9
#> 3   8   8   8   3   4   8   5   3   3   1   1   2   6   2   8  10   6  10   4
#> 4  10   5   1   8   4   5   4   4   4   6   4   7   2   8   4   7   7   1   4
#> 5   1   8   5  10   2   4   4   1   9   2   5   8   8   2   8   1   3   5   3
#>   TGG TGT TTA TTC TTG TTT
#> 1   2   5   3   8   6  10
#> 2   8   3   5  10   7   9
#> 3   4  10   2   1   7   5
#> 4   4   5   4   5  10  10
#> 5   7   2   1   2   7   1

cT <- setKO(cT, rep(c("K00001", "K00002"), 5))
subset(cT, "K00001")

#> codonTable instance with codon counts from 5 sequences.
#>   length     KO AAA AAC AAG AAT ACA ACC ACG ACT AGA AGC AGG AGT ATA ATC ATG ATT
#> 1    357 K00001  10   9  10   4   1   2   8   7   8   9   9   1   6   3   5   8
#> 2    327 K00001   3   1   2   3   9   1   3   8   8   9   5   8   5   4   2   1
#> 3    299 K00001   9   8   2   3   1   3   3   2   6   8   7   1   1   5   6   3
#> 4    344 K00001   2   7   9   1   2   9   8   8   8   5   8  10   7   2   7  10
#> 5    307 K00001   4   5   1   1   7   3   6   6   9   2   6  10   9   7  10   9
#>   CAA CAC CAG CAT CCA CCC CCG CCT CGA CGC CGG CGT CTA CTC CTG CTT GAA GAC GAG
#> 1   4  10   6   3   5   4   5   4   6   2   4   9   4   1  10   4   9   9   5
#> 2   5   5   1   7   1   3   8   2   9   1   4  10   7   7   5   6   3   4   9
#> 3   4   5   3   1   8   1  10   9   3   8   8   7   3   1   7   6   1   2   7
#> 4   3   4   7   4   4  10   8   2   2   9   3   3   4   9   2   3  10   8   3
#> 5   5   3   1  10   1   5   2   9   2   8   7   4   7   2   5   1  10   7   6
#>   GAT GCA GCC GCG GCT GGA GGC GGG GGT GTA GTC GTG GTT TAC TAT TCA TCC TCG TCT
#> 1   2   9  10   7   2   4   1   3   2   8   5   9  10   9   5   5   8   4   6
#> 2   2   8   3   1   6   7  10   5   9   7   1   3   7   8   5   6   1  10   8
#> 3   8   8   8   8   3   4   8   5   3   3   1   1   2   6   2   8  10   6  10
#> 4  10  10   5   1   8   4   5   4   4   4   6   4   7   2   8   4   7   7   1
#> 5   8   1   8   5  10   2   4   4   1   9   2   5   8   8   2   8   1   3   5
#>   TGC TGG TGT TTA TTC TTG TTT
#> 1  10   2   5   3   8   6  10
#> 2   9   8   3   5  10   7   9
#> 3   4   4  10   2   1   7   5
#> 4   4   4   5   4   5  10  10
#> 5   3   7   2   1   2   7   1

cT <- setCOG(cT, rep(c("COG0001", "COG0002"), 5))
subset(cT, "COG0001")

#> codonTable instance with codon counts from 5 sequences.
#>   length     COG     KO AAA AAC AAG AAT ACA ACC ACG ACT AGA AGC AGG AGT ATA ATC
#> 1    357 COG0001 K00001  10   9  10   4   1   2   8   7   8   9   9   1   6   3
#> 2    327 COG0001 K00001   3   1   2   3   9   1   3   8   8   9   5   8   5   4
#> 3    299 COG0001 K00001   9   8   2   3   1   3   3   2   6   8   7   1   1   5
#> 4    344 COG0001 K00001   2   7   9   1   2   9   8   8   8   5   8  10   7   2
#> 5    307 COG0001 K00001   4   5   1   1   7   3   6   6   9   2   6  10   9   7
#>   ATG ATT CAA CAC CAG CAT CCA CCC CCG CCT CGA CGC CGG CGT CTA CTC CTG CTT GAA
#> 1   5   8   4  10   6   3   5   4   5   4   6   2   4   9   4   1  10   4   9
#> 2   2   1   5   5   1   7   1   3   8   2   9   1   4  10   7   7   5   6   3
#> 3   6   3   4   5   3   1   8   1  10   9   3   8   8   7   3   1   7   6   1
#> 4   7  10   3   4   7   4   4  10   8   2   2   9   3   3   4   9   2   3  10
#> 5  10   9   5   3   1  10   1   5   2   9   2   8   7   4   7   2   5   1  10
#>   GAC GAG GAT GCA GCC GCG GCT GGA GGC GGG GGT GTA GTC GTG GTT TAC TAT TCA TCC
#> 1   9   5   2   9  10   7   2   4   1   3   2   8   5   9  10   9   5   5   8
#> 2   4   9   2   8   3   1   6   7  10   5   9   7   1   3   7   8   5   6   1
#> 3   2   7   8   8   8   8   3   4   8   5   3   3   1   1   2   6   2   8  10
#> 4   8   3  10  10   5   1   8   4   5   4   4   4   6   4   7   2   8   4   7
#> 5   7   6   8   1   8   5  10   2   4   4   1   9   2   5   8   8   2   8   1
#>   TCG TCT TGC TGG TGT TTA TTC TTG TTT
#> 1   4   6  10   2   5   3   8   6  10
#> 2  10   8   9   8   3   5  10   7   9
#> 3   6  10   4   4  10   2   1   7   5
#> 4   7   1   4   4   5   4   5  10  10
#> 5   3   5   3   7   2   1   2   7   1